DTW/App/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
eventlet>=0.35.0
python-pptx>=1.0.0
python-docx>=1.1.0
boto3>=1.34.0
